Fix OrderCustomization to correctly set the CallOffId Customization and pull in some of the other settings from ORDAPI Reorder the Customizations in Fixture Factory due to their ordering nature. Move GenerateUrlFromMethod to the Test base Add Version of NavigateToUrl that automatically called the GenerateUrlFromMethod Update NavigateToUrl in Supplier Information tests to use new NavigateToUrl Fix some of the Broken Unit Tests Update the FormatStringForComparision to strip out all whitespace. Updat some E2E tests to use the FormatStringForComparision Fix Capabilities to not use hardcoded number of solutions values Fix Manage Catalogue Solutions to not filter by Supplier
