+++ tests/NHSD.GPIT.BuyingCatalogue.E2E.PublicBrowseTests/Areas/Admin/AddNewSolution/DevelopmentPlans.cs
@@ -24,7 +24,9 @@ namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Admin.AddNewSolution
             await using var context = GetEndToEndDbContext();
             var solutionName = (await context.CatalogueItems.SingleAsync(s => s.CatalogueItemId == new CatalogueItemId(99999, "888"))).Name;
 
-            CommonActions.PageTitle().Should().BeEquivalentTo($"Development plans - {solutionName}");
+            CommonActions.PageTitle()
+                .Should()
+                .BeEquivalentTo(CommonActions.FormatStringForComparison($"Development plans - {solutionName}"));
         }
 
         [Fact]
