+++ tests/NHSD.GPIT.BuyingCatalogue.Test.Framework/AutoFixtureCustomisations/CommonAutoDataAttribute.cs
@@ -12,8 +12,8 @@ namespace NHSD.GPIT.BuyingCatalogue.Test.Framework.AutoFixtureCustomisations
                     new AutoMoqCustomization(),
                     new AddSolutionModelCustomization(),
                     new CreateSolutionModelCustomization(),
-                    new OrderCustomization(),
                     new CallOffIdCustomization(),
+                    new OrderCustomization(),
                     new CatalogueItemIdCustomization(),
                     new ControllerBaseCustomization(),
                     new IgnoreCircularReferenceCustomisation(),
