+++ tests/NHSD.GPIT.BuyingCatalogue.Test.Framework/AutoFixtureCustomisations/OrderCustomization.cs
@@ -1,4 +1,5 @@
 ﻿using AutoFixture;
+using AutoFixture.Kernel;
 using NHSD.GPIT.BuyingCatalogue.EntityFramework.Ordering.Models;
 
 namespace NHSD.GPIT.BuyingCatalogue.Test.Framework.AutoFixtureCustomisations
@@ -7,11 +8,15 @@ namespace NHSD.GPIT.BuyingCatalogue.Test.Framework.AutoFixtureCustomisations
     {
         public void Customize(IFixture fixture)
         {
-            var callOffId = fixture.Create<CallOffId>();
+            var callOffId = fixture
+                .Customize(new CallOffIdCustomization())
+                .Create<CallOffId>();
 
             fixture.Customize<Order>(order => order
             .With(o => o.Id, callOffId.Id)
-            .With(o => o.Revision, callOffId.Revision));
+            .With(o => o.Revision, callOffId.Revision)
+            .Without(o => o.IsDeleted)
+            .Without(o => o.OrderStatus));
         }
     }
 }
